+++ src/chartdataset_p.h
@@ -53,7 +53,7 @@ public:
 
     void zoomInDomain(const QRectF& rect, const QSizeF& size);
     void zoomOutDomain(const QRectF& rect, const QSizeF& size);
-    void scrollDomain(int dx,int dy,const QSizeF& size);
+    void scrollDomain(qreal dx,qreal dy,const QSizeF& size);
 
     int seriesCount(QAbstractSeries::SeriesType type);
     int seriesIndex(QAbstractSeries *series);
